fix: do not set unwriteable fields
A common pattern for use of this module is to catch an error from a
third party module or API and add a `.code` field to it.

If the caught error already has a `.code` field and that field has
been defined as read-only, this will fail with a `TypeError` that
masks the underlying problem.

The change here is to make a best-effort attempt to assign the `.code`
field and other props, but to respect the orignal value if something
has been set and marked read-only.