fix: create a new error object when assigning code/props fails
This module is supposed to add a `.code` property to an `Error` object
- if this fails, it can create hard to track down bugs as the calling
code expects there to be a `.code` property on the error.

Here we create a new `Error` if assigning the passed `.code` or props
fails. The original error message and stack trace are used on the new
error, and the original error instance is accessible via the `.cause`
property on the returned error.