fix: remove instanceof check for passed error (#26)
When an Error is created in a different context (e.g. in a browser
iframe or similar), `instanceof` doesn't work.

The change here is to remove the `instanceof` check and instead
just ensure that the `err` argument is truthy and is not a string
to catch the older style of passing an error message in as the first
argument.